Automatically determine whether to test with or without a trailing slash
This changes the behavior of this module to be smarter about whether it
should test the new URL path with or without a trailing slash. Since
this middleware should come after `app.router`, we already know that the
URL path as is does not have a matching route. If that original path
contains a slash, then it should be remove and visa versa. This new URL
path can then be checked against the app's routes to determine if there
is a match.

Closes #1